SDK-adapters merge refactoring (#495)
* SDK-adapters merge refactoring - first cut

* Refactor sdk back to unstract.sdk

* Upgrade to SDK 0.39.0
